@@ -3,7 +3,7 @@ package io.mosip.authentication.esignet.integration.dto;
 import java.util.List;
 import java.util.Map;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 @Data
